requirements.txt
django

django-bootstrap4

django-cors-headers

psycopg2-binary

# Code checking
pycodestyle
pylint
pylint-django

